backend/app/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
neo4j>=5.14
sentence-transformers==2.2.2
torch==2.2.0
requests==2.31.0
openai
python-dotenv
reportlab
